(BKR-623) Refactor Structure and Outcomes modules
