+++ acceptance/tests/base/dsl/structure_test.rb
@@ -8,7 +8,7 @@ test_name "dsl::structure" do
 
       assert_equal 1, @in_confine, "#confine_block did not run the supplied block"
 
-    rescue Beaker::DSL::Outcomes::SkipTest => e
+    rescue Beaker::DSL::SkipTest => e
       fail "#confine_block raised unexpected SkipTest exception: #{e}"
     end
   end
@@ -25,7 +25,7 @@ test_name "dsl::structure" do
       assert_equal 1, @in_confine, "#confine_block did not run the supplied block"
       assert_equal hosts.dup, hosts, "#confine_block did not preserve the hosts array"
 
-    rescue Beaker::DSL::Outcomes::SkipTest => e
+    rescue Beaker::DSL::SkipTest => e
       fail "#confine_block raised unexpected SkipTest exception: #{e}"
     end
   end
@@ -39,7 +39,7 @@ test_name "dsl::structure" do
 
       assert_equal 0, @in_confine, "#confine_block did not skip the supplied block"
 
-    rescue Beaker::DSL::Outcomes::SkipTest => e
+    rescue Beaker::DSL::SkipTest => e
       fail "#confine_block raised unexpected SkipTest exception: #{e}"
     end
   end
@@ -56,7 +56,7 @@ test_name "dsl::structure" do
       assert_equal 0, @in_confine, "#confine_block did not skip the supplied block"
       assert_equal hosts.dup, hosts, "#confine_block did not preserve the hosts array"
 
-    rescue Beaker::DSL::Outcomes::SkipTest => e
+    rescue Beaker::DSL::SkipTest => e
       fail "#confine_block raised unexpected SkipTest exception: #{e}"
     end
   end
@@ -70,7 +70,7 @@ test_name "dsl::structure" do
         @in_confine +=1
         skip_test "this block raises a skip"
       end
-    rescue Beaker::DSL::Outcomes::SkipTest => e
+    rescue Beaker::DSL::SkipTest => e
       assert_match /this block raises a skip/, e.message, "#confine_block raised an unexpected skip_test"
       assert_equal 1, @in_confine, "#confine_block did not execute supplied block"
       assert_equal hosts.dup, hosts, "#confine_block did not preserve the hosts array"
