+++ lib/beaker/dsl.rb
@@ -1,5 +1,4 @@
-[ 'install_utils', 'roles', 'outcomes', 'assertions', 'patterns',
-  'structure', 'helpers', 'wrappers' ].each do |lib|
+[ 'install_utils', 'roles', 'assertions', 'patterns', 'helpers', 'wrappers' ].each do |lib|
   require "beaker/dsl/#{lib}"
 end
 
@@ -71,13 +70,32 @@ module Beaker
   #
   #@api dsl
   module DSL
+
+    # Raise this class if it is determined that a test case should not
+    # be executed because the feature in question is still a
+    # "Work in Progress"
+    class PendingTest < Exception; end
+
+    # Raise this class if execution should be stopped because the test
+    # is not applicable within a given environment.
+    class SkipTest    < Exception; end
+
+    # Raise this class if some criteria has been met that proves a failure.
+    class FailTest    < Exception; end
+
+    # Raise this class if execution should stop because enough criteria has
+    # shown itself to pass the test.
+    class PassTest    < Exception; end
+
     include Beaker::DSL::Roles
-    include Beaker::DSL::Outcomes
-    include Beaker::DSL::Structure
     include Beaker::DSL::Assertions
     include Beaker::DSL::Wrappers
     include Beaker::DSL::Helpers
     include Beaker::DSL::InstallUtils
     include Beaker::DSL::Patterns
+
+    def self.register(helper_module)
+      include helper_module
+    end
   end
 end
