+++ spec/beaker/dsl/install_utils/pe_utils_spec.rb
@@ -1,12 +1,7 @@
 require 'spec_helper'
 
 class ClassMixedWithDSLInstallUtils
-  include Beaker::DSL::InstallUtils
-  include Beaker::DSL::Wrappers
-  include Beaker::DSL::Helpers
-  include Beaker::DSL::Structure
-  include Beaker::DSL::Roles
-  include Beaker::DSL::Patterns
+  include Beaker::DSL
 
   attr_accessor :hosts
